Refactor for contao 4.13 and PHP 8
